refactor(reef): restructure schema explorer around routes (review feedback)
Address Ludo's review:
- Split views/schema-explorer.tsx into a views/schema-explorer/ folder
  mirroring the three routes (schema, schema-table, schema-empty), with
  shared helpers in shared.ts.
- Routes are now thin re-exports: loader + default component + ErrorBoundary
  come from the views.
- Error handling moved from <Await errorElement> to route ErrorBoundaries:
  the deferred promises are unwrapped with React's use(), so a rejection
  propagates to the boundary — the same path a non-deferred loader error
  would take. Schema failure errors the whole page (critical data); a
  columns failure replaces only the detail panel. Retry re-runs the
  navigation (revalidation alone does not reset a render-error boundary).
- Drop redundant as="span" on Typography (span is the default).

Held for Ludo's stack (SSR #1568 + nav progress bar #1543): switching
clientLoader to a server loader and un-deferring the schema fetch so the
global progress bar is the top-level pending UI.
