+++ apps/reef/app/routes/schema.tsx
@@ -1,14 +1,14 @@
 import { fetchSchemaFromCoral } from '@/lib/schema-explorer'
-import { SchemaExplorer } from '@/views/schema-explorer'
+import { SchemaExplorer, SchemaExplorerError } from '@/views/schema-explorer/schema'
 
 import type { Route } from './+types/schema'
 
-// Desktop is SPA-only (no server) and the catalog is fetched over gRPC-web via
-// the browser transport + desktop bridge, so this must be a clientLoader. The
-// schema promise is deferred (not awaited) so the route renders immediately and
-// the view streams it in with Suspense — SPA mode forbids a route HydrateFallback.
-// The route's abort signal cancels the catalog request when the navigation is
-// superseded instead of leaving it running against the sidecar.
+// The catalog is fetched over gRPC-web via the browser transport + desktop
+// bridge, so this must be a clientLoader (until the SSR/BFF stack lands and a
+// server loader can talk to the sidecar directly). The schema promise is
+// deferred (not awaited) so the route renders immediately and the view streams
+// it in with Suspense — SPA mode forbids a route HydrateFallback. The abort
+// signal cancels the catalog request when the navigation is superseded.
 export function clientLoader({ request }: Route.ClientLoaderArgs) {
   return { schema: fetchSchemaFromCoral(request.signal) }
 }
@@ -18,3 +18,5 @@ clientLoader.hydrate = true as const
 export default function SchemaRoute({ loaderData }: Route.ComponentProps) {
   return <SchemaExplorer schema={loaderData.schema} />
 }
+
+export { SchemaExplorerError as ErrorBoundary }
